requirements.txt
streamlit>=1.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pandas>=1.3.0
tensorflow>=2.5.0
keras
gtts>=2.2.3
Pillow>=8.3.0
plotly>=5.3.0
# For optional audio playback (if using helper function)
playsound>=1.2.2 
pygame>=2.0.1
# For optional voice commands/speech recognition
SpeechRecognition>=3.8.1
PyAudio
